processNonHistogramLikeMetircs: ignore unsupported types
Encountered in the wild, a user was getting Prometheus metrics from
cadvisor but unfortunately their scrape featured "summary" which
isn't yet supported in the OpenCensus API (at least to the least
of my knowledge). Due to the unwiedly number of types of metrics
out there, it is prudent to log the unsupported type but return
a nil error so that processing can go on per scrape.

The log will look something like this:

    error="unknown recognized metric type" type=summary

Fixes #3